2.5: Separate module step 10
